Fix bugs generating TwiML response messages
This change corrects a series of bugs in how the TwiML response strings
are generated. While the TwiML was correctly generated in some cases,
there were issues when a response was either not received or an
unexpected response was received.

In addition, the change introduces covering tests to help ensure the
correctness and quality of the code.